+++ aten/src/ATen/core/dispatch/DispatchTable.h
@@ -1,12 +1,15 @@
 #pragma once
 
 #include <ATen/core/function_schema.h>
+#include <ATen/core/function_schema_argument.h>
 #include <c10/util/LeftRight.h>
 #include <c10/util/Metaprogramming.h>
 #include <c10/util/flat_hash_map.h>
 #include <c10/util/either.h>
 #include <ATen/core/ivalue.h>
 #include <ATen/core/dispatch/KernelFunction.h>
+#include <ATen/core/stack.h>
+#include <ATen/core/function_schema.h>
 
 #include <array>
 #include <atomic>
@@ -19,13 +22,6 @@
 
 namespace c10 {
 
-/**
- * The type of a user-supplied function to initialize the kernel cache.
- * this is stored together with the KernelFunction in the DispatchTable
- * so we can create a new cache instance when a kernel is looked up
- * from the dispatch table.
- */
-using KernelCacheCreatorFunction = std::function<std::unique_ptr<c10::KernelCache> ()>;
 /**
  * The dispatch table stores a pointer to a kernel function and a pointer
  * to a function initializing a cache for the kernel. If the kernel wants
