+++ aten/src/ATen/core/dispatch/OperatorEntry.cpp
@@ -1,3 +1,4 @@
+#include <ATen/core/dispatch/Dispatcher.h>
 #include <ATen/core/dispatch/OperatorEntry.h>
 
 namespace c10 {
