+++ aten/src/ATen/core/function_schema.h
@@ -1,122 +1,22 @@
 #pragma once
 
-#include <ATen/core/jit_type.h>
 #include <ATen/core/interned_strings.h>
-#include <ATen/core/ivalue.h>
 #include <ATen/core/alias_info.h>
+#include <ATen/core/operator_name.h>
+#include <c10/util/Optional.h>
 #include <unordered_map>
+#include <memory>
 
 namespace c10 {
+class IValue;
+struct Type;
+using TypePtr = std::shared_ptr<Type>;
 
 // schema as used in the compiler for resolving function calls and reporting
 // errors. These objects should be constructed from C10 schema once those
 // are available.
 
-struct Argument {
-  Argument(
-      std::string name = "",
-      TypePtr type = nullptr,
-      c10::optional<int32_t> N = c10::nullopt,
-      c10::optional<IValue> default_value = c10::nullopt,
-      bool kwarg_only = false,
-      c10::optional<AliasInfo> alias_info = c10::nullopt,
-      bool is_inferred_type = false)
-      : name_(std::move(name)),
-        type_(type ? type : TensorType::get()),
-        N_(std::move(N)),
-        default_value_(std::move(default_value)),
-        kwarg_only_(kwarg_only),
-        alias_info_(std::move(alias_info)),
-        is_inferred_type_(is_inferred_type) {
-    if (default_value_ && default_value_->isTensor()) {
-      auto t = default_value_->toTensor();
-      AT_ASSERT(!t.defined() || t.is_variable());
-    }
-  }
-  const std::string& name() const {
-    return name_;
-  }
-  TypePtr type() const {
-    return type_;
-  }
-  c10::optional<int32_t> N() const {
-    return N_;
-  }
-  const c10::optional<IValue>& default_value() const {
-    return default_value_;
-  }
-  bool kwarg_only() const {
-    return kwarg_only_;
-  }
-  const c10::optional<AliasInfo>& alias_info() const {
-    return alias_info_;
-  }
-  bool is_inferred_type() const {
-    return is_inferred_type_;
-  }
-  std::string formatTypeMismatchMsg(const std::string& actual_type) const {
-    std::string inferred_type_hint;
-    if (is_inferred_type()) {
-      inferred_type_hint = c10::str(
-          "Inferred '",
-          name(),
-          "' to be of type 'Tensor' ",
-          "because it was not annotated with an explicit type.\n");
-    }
-    return c10::str(
-        "Expected a value of type '",
-        type()->python_str(),
-        "' for argument '",
-        name(),
-        "' but instead found type '",
-        actual_type,
-        "'.\n",
-        inferred_type_hint);
-  }
-
-  Argument cloneWithType(TypePtr new_type) const {
-    return Argument(name_, new_type, N_, default_value_, kwarg_only_, alias_info_);
-  }
-
-private:
-  std::string name_;
-  TypePtr type_;
-  // for list types, an optional statically known length for the list
-  // e.g. for int[3]: type = ListType::ofInts(), N = 3
-  // If present, this will allow scalars to be broadcast to this length to
-  // become a list.
-  c10::optional<int32_t> N_;
-
-  c10::optional<IValue> default_value_;
-  // is this only specifyable as a keyword argument?
-  bool kwarg_only_;
-  c10::optional<AliasInfo> alias_info_;
-  bool is_inferred_type_;
-};
-
-namespace detail {
-inline bool defaultValueEquals_(const c10::optional<IValue>& lhs, const c10::optional<IValue>& rhs) {
-  if (lhs.has_value()) {
-    return rhs.has_value() && impl::shallowEquals(*lhs, *rhs);
-  } else {
-    return !rhs.has_value();
-  }
-}
-}
-
-inline bool operator==(const Argument& lhs, const Argument& rhs) {
-  return lhs.name() == rhs.name()
-          && *lhs.type() == *rhs.type()
-          && lhs.N() == rhs.N()
-          && detail::defaultValueEquals_(lhs.default_value(), rhs.default_value())
-          && lhs.kwarg_only() == rhs.kwarg_only()
-          && lhs.alias_info() == rhs.alias_info();
-}
-
-struct OperatorName final {
-  std::string name;
-  std::string overload_name;
-};
+struct Argument;
 
 struct FunctionSchema {
   FunctionSchema(
@@ -181,30 +81,10 @@ public:
   bool is_varret() const {
     return is_varret_;
   }
-  bool is_mutable() const {
-    return std::any_of(
-        arguments_.cbegin(), arguments_.cend(), [](const Argument& arg) {
-          const auto& aliasInfo = arg.alias_info();
-          return aliasInfo && aliasInfo.value().isWrite();
-        });
-  }
+  bool is_mutable() const;
 
-  c10::optional<int> argumentIndexWithName(const std::string& name) const {
-    for(size_t i = 0; i < arguments().size(); ++i) {
-      if(name == arguments()[i].name())
-        return i;
-    }
-    return c10::nullopt;
-  }
-  FunctionSchema cloneWithArguments(std::vector<Argument> new_arguments) const {
-    return FunctionSchema(
-        name(),
-        overload_name(),
-        std::move(new_arguments),
-        returns(),
-        is_vararg(),
-        is_varret());
-  }
+  c10::optional<int> argumentIndexWithName(const std::string& name) const;
+  FunctionSchema cloneWithArguments(std::vector<Argument> new_arguments) const;
 
   std::string formatTypeMismatchMsg(
       const Argument& expected,
@@ -223,82 +103,17 @@ public:
 
   void findErrorInKwargs(const std::vector<std::string>& kwargs) const;
 
-  bool hasAnyAliasInfo() const {
-    for (const auto& arg : arguments_) {
-      if (arg.alias_info().has_value()) {
-        return true;
-      }
-    }
-    for (const auto& ret : returns_) {
-      if (ret.alias_info().has_value()) {
-        return true;
-      }
-    }
-    return false;
-  }
+  bool hasAnyAliasInfo() const;
 };
 
-inline bool operator==(const FunctionSchema& lhs, const FunctionSchema& rhs) {
-  return lhs.name() == rhs.name()
-      && lhs.overload_name() == rhs.overload_name()
-      && lhs.arguments() == rhs.arguments()
-      && lhs.returns() == rhs.returns()
-      && lhs.is_vararg() == rhs.is_vararg()
-      && lhs.is_varret() == rhs.is_varret();
-}
-
-inline bool operator!=(const FunctionSchema& lhs, const FunctionSchema& rhs) {
-  return !(lhs == rhs);
-}
+bool operator==(const FunctionSchema& lhs, const FunctionSchema& rhs);
+bool operator!=(const FunctionSchema& lhs, const FunctionSchema& rhs);
 
 // print out Argument, which is compatible with FunctionSchema parser
 // full format: Type(alias)? name=default_value
-inline std::ostream& operator<<(std::ostream& out, const Argument& arg) {
-  bool optional_type = arg.type()->isSubclass(TypeKind::OptionalType);
-  // for adjusting the ? position.
-  // in schema, we have Tensor?(a!) input, and t(a!)?.
-  // however, t?(a!) doesn't work with schema parser.
-  // so we always use Type(alias)? format
-  std::stringstream oss;
-  if (arg.type()->isSubclass(TypeKind::ListType) && arg.N()) {
-    oss << arg.type()->cast<ListType>()->getElementType()->str();
-    oss << "[" << arg.N().value() << "]";
-  } else {
-    oss << arg.type()->str();
-  }
-  if (optional_type) {
-    oss.seekp(oss.str().size() - 1);
-  }
-  if (arg.alias_info()) {
-    oss << arg.alias_info().value();
-  }
-  if (optional_type) {
-    oss << "?";
-  }
-  out << oss.str();
-  if (!arg.name().empty()) {
-    out << " " << arg.name();
-  }
-  if (arg.default_value()) {
-    out << "=";
-    if (arg.type()->kind() == c10::TypeKind::StringType) {
-        // TODO prettify the result, such as using \n to represent \012
-        out << "\'";
-        std::ios_base::fmtflags flags(out.flags());
-        for (unsigned char c : arg.default_value().value().toStringRef()) {
-          out << "\\" << std::oct << std::setfill('0') << std::setw(3)
-            << static_cast<uint64_t>(c);
-        }
-        out.flags(flags);
-        out << "\'";
-    } else {
-      out << arg.default_value().value();
-    }
-  }
-  return out;
-}
+std::ostream& operator<<(std::ostream& out, const Argument& arg);
 
-inline std::ostream& operator<<(std::ostream& out, const FunctionSchema& schema);
+std::ostream& operator<<(std::ostream& out, const FunctionSchema& schema);
 
 inline std::string toString(const FunctionSchema& schema) {
   std::ostringstream str;
