+++ aten/src/ATen/core/ivalue.h
@@ -2,8 +2,16 @@
 
 #include <ATen/core/blob.h>
 #include <c10/util/intrusive_ptr.h>
-#include <ATen/core/Tensor.h>
+#include <c10/util/Optional.h>
+#include <c10/core/Scalar.h>
+#include <c10/core/Device.h>
+#include <c10/core/Layout.h>
+#include <c10/core/MemoryFormat.h>
+#include <c10/core/QScheme.h>
 
+namespace at {
+class Tensor;
+}
 namespace torch {
 namespace jit {
 struct Function;
@@ -17,6 +25,7 @@ template<class Key, class Value> class Dict;
 template<class T> class List;
 struct IValue;
 struct ClassType;
+class TensorImpl;
 namespace ivalue {
 struct Tuple;
 struct Future;
@@ -79,29 +88,8 @@ struct CAFFE2_API IValue final {
 
   void dump() const;
 
-  bool isAliasOf(const IValue& rhs) const {
-    if (this->tag != rhs.tag) {
-      // Trivially don't alias if the type is different
-      return false;
-    }
+  bool isAliasOf(const IValue& rhs) const;
 
-    if (!this->is_intrusive_ptr) {
-      // Primitive types don't alias anything
-      return false;
-    }
-
-    AT_ASSERT(rhs.is_intrusive_ptr);
-
-    // Tensors should be compared based on internal storage
-    if (this->isTensor()) {
-      const auto thisTensor = this->toTensor();
-      const auto rhsTensor = rhs.toTensor();
-      return thisTensor.is_alias_of(rhsTensor);
-    }
-
-    // Other types can be compared by their ptr value
-    return this->payload.as_intrusive_ptr == rhs.payload.as_intrusive_ptr;
-  }
   void swap(IValue & rhs) noexcept {
     std::swap(payload, rhs.payload);
     std::swap(is_intrusive_ptr, rhs.is_intrusive_ptr);
@@ -113,21 +101,12 @@ struct CAFFE2_API IValue final {
   // we prefer to write them manually for clarity
 
   // Tensor
-  IValue(at::Tensor t)
-  : tag(Tag::Tensor), is_intrusive_ptr(t.defined())  {
-    // Note: the undefined tensor is not refcounted, so while it
-    // is tagged as a tensor, is_intrusive_ptr is set to false.
-    // This is not an optional optimization: our incref call
-    // *will not* do the right thing when called on an
-    // undefined tensor.
-    payload.as_intrusive_ptr = t.unsafeReleaseTensorImpl();
-  }
+  IValue(at::Tensor&& t);
+  IValue(const at::Tensor& t);
   bool isTensor() const { return Tag::Tensor == tag; }
   at::Tensor toTensor() &&;
   at::Tensor toTensor() const &;
-  at::TensorImpl* unsafeToTensorImpl() const {
-    return static_cast<at::TensorImpl*>(payload.as_intrusive_ptr);
-  }
+  at::TensorImpl* unsafeToTensorImpl() const;
 
   const IValue& toIValue() const {
     return *this;
