+++ aten/src/ATen/core/jit_type.h
@@ -1,18 +1,20 @@
 #pragma once
 
-#include <ATen/core/Tensor.h>
 #include <ATen/core/functional.h>
 #include <ATen/core/interned_strings.h>
-#include <ATen/core/ivalue.h>
 #include <ATen/core/qualified_name.h>
 #include <c10/util/TypeList.h>
-
 #include <c10/util/Optional.h>
+#include <c10/core/ScalarType.h>
+#include <c10/core/Device.h>
 
 #include <iostream>
 #include <memory>
 #include <type_traits>
 
+namespace at {
+class Tensor;
+}
 namespace torch {
 namespace jit {
 struct Function;
@@ -23,6 +25,10 @@ struct CompilationUnit;
 } // namespace torch
 
 namespace c10 {
+class IValue;
+class Scalar;
+template<class T> class List;
+template<class K, class V> class Dict;
 
 struct FunctionSchema;
 using OptNameList = c10::optional<std::vector<std::string>>;
@@ -418,13 +424,7 @@ struct CAFFE2_API DimensionedTensorType : public TensorType {
  protected:
   DimensionedTensorType(
       const at::Tensor& tensor,
-      TypeKind kind = TypeKind::DimensionedTensorType)
-      : DimensionedTensorType(
-            tensor.scalar_type(),
-            tensor.device(),
-            tensor.dim(),
-            tensor.is_variable() && tensor.requires_grad(),
-            kind) {}
+      TypeKind kind = TypeKind::DimensionedTensorType);
   DimensionedTensorType(
       at::ScalarType scalar_type,
       at::Device device,
@@ -624,13 +624,7 @@ struct CAFFE2_API ProfiledTensorType : public TensorType {
   static const TypeKind Kind = TypeKind::ProfiledTensorType;
 
  private:
-  ProfiledTensorType(const at::Tensor& tensor)
-      : TensorType(),
-        scalar_type_(tensor.scalar_type()),
-        device_(tensor.device()),
-        sizes_(tensor.sizes().vec()),
-        strides_(tensor.strides().vec()),
-        requires_grad_(tensor.requires_grad()) {}
+  ProfiledTensorType(const at::Tensor& tensor);
   ProfiledTensorType(
       c10::optional<at::ScalarType> scalar_type,
       c10::optional<Device> device,
@@ -746,10 +740,7 @@ struct CAFFE2_API CompleteTensorType : public DimensionedTensorType {
   static TypePtr fromBoolType();
 
  private:
-  CompleteTensorType(const at::Tensor& tensor)
-      : DimensionedTensorType(tensor, TypeKind::CompleteTensorType),
-        sizes_(tensor.sizes().vec()),
-        strides_(tensor.strides().vec()) {}
+  CompleteTensorType(const at::Tensor& tensor);
   CompleteTensorType(
       at::ScalarType scalar_type,
       at::Device device,
@@ -1633,3 +1624,5 @@ struct CAFFE2_API ClassType : public NamedType {
   std::vector<Function*> methods_;
 };
 } // namespace c10
+
+#include "jit_type_inl.h"
