+++ c10/util/LeftRight.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <atomic>
 #include <functional>
 #include <mutex>
